uint8 => uint256

+++ contracts/egg.sol
@@ -131,10 +131,10 @@ contract RoosterEgg is ERC721, ERC721Burnable, Ownable, Pausable {
     require(amount + purchasedAmount[purchaser] <= cap, "Exceeds cap");
   }
 
-  function _mintEggs(address to, uint8 amount) private {
+  function _mintEggs(address to, uint256 amount) private {
     uint24 newtokenId = _tokenIdCounter;
 
-    for (uint8 i = 0; i < amount; i++) {
+    for (uint256 i = 0; i < amount; i++) {
       _safeMint(to, newtokenId++);
     }
 
@@ -170,7 +170,7 @@ contract RoosterEgg is ERC721, ERC721Burnable, Ownable, Pausable {
     emit NewPresale(supply_, cap_, openingTime_, closingTime_, price_, cashbackPerEgg_);
   }
 
-  function mintEggs(address to, uint8 amount) external onlyOwner {
+  function mintEggs(address to, uint256 amount) external onlyOwner {
     _mintEggs(to, amount);
   }
